+++ unicode_to_png.py
@@ -3,7 +3,7 @@
 # If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/
 #
 # Original Author: Sergio Palma Hidalgo
-# Project URL: https://github.com/sergiopalmah/unicode_to_png
+# Project URL: https://github.com/del-Pacifico/unicode-to-png
 # Copyright (c) 2025 Sergio Palma Hidalgo
 # All rights reserved.
 #